Patch (rat-64bit.patch) from Doug 20 july 2007 fixes - thanks: "With 64bit systems, the RAT UI would show ghost ssrc entries as a 64bit value, the upper 32bits were random values, but the lower 32bits corresponded to existing entries."

+++ ui_send_audio.c
@@ -606,7 +606,7 @@ ui_send_audio_3d_settings(session_t *sp, char *addr, uint32_t ssrc)
 
         render_3D_get_parameters(p->render_3D_data, &azimuth, &filter_type, &filter_length);
         filter_name = mbus_encode_str(render_3D_filter_get_name(filter_type));
-        mbus_qmsgf(sp->mbus_engine, addr, TRUE, "audio.3d.user.settings", "\"%08lx\" %s %d %d", ssrc, filter_name, filter_length, azimuth);
+        mbus_qmsgf(sp->mbus_engine, addr, TRUE, "audio.3d.user.settings", "\"%08lx\" %s %d %d", (unsigned long) ssrc, filter_name, filter_length, azimuth);
         xfree(filter_name);
 }
